requirements.txt
pdfplumber
PyPDF2
pdfminer.six
PyMuPDF
# openai==0.27.9
Pillow
